fix(cmd-bar): Update CmdBarFilter slot options
The existing CmdBarFilter slot options are updated to include 'hidden-options'. The 'option' slot properties in the CmdBarFilter.vue have been modified to take an array of 'hiddenOptions'. This change will help to extend the functionality and versatility of the component, to meet the demands of any potential hidden features within the FilterOption.